For #6557 - split base extension functionality from ads
